[#319][ADD] Add frontend_homepage.
