+++ main/local_modules/frontend/templates/website_partner.xml
@@ -26,14 +26,6 @@
             <field name="parent_id" ref="frontend.directory"/>
             <field name="sequence" type="int">10</field>
         </record>
-        <record id="dashboard" model="website.menu">
-            <field name="name">Directory</field>
-            <field name="url">/web/</field>
-            <field name="parent_id" ref="website.main_menu"/>
-            <field name="sequence" type="int">10</field>
-            <field name="user_logged" eval="True"/>
-            <field name="user_not_logged" eval="False"/>
-        </record>
 
         <template id="frontend.partner" name="Frontend: partner" page="True">
             <t t-call="website.layout">
