+++ main/local_modules/frontend_homepage/README.rst
@@ -1,0 +1,16 @@
+Frontend Homepage
+=================
+Module dedicated to the generation of the homepage(s) of the website.
+
+It should be composed by
+* templates
+* models
+* controllers
+
+needed by the page.
+
+
+Contributors
+------------
+* Jordi Riera <kender.jr@gmail.com>
+* David Mazeau <d.mazeau@gmail.com>
