+++ main/local_modules/frontend_homepage/data/website_menus.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <record id="dashboard" model="website.menu">
+            <field name="name">Directory</field>
+            <field name="url">/web</field>
+            <field name="parent_id" ref="website.main_menu"/>
+            <field name="sequence" type="int">10</field>
+            <field name="user_logged" eval="True"/>
+            <field name="user_not_logged" eval="False"/>
+        </record>
+    </data>
+</openerp>
