+++ main/local_modules/frontend_homepage/templates/template_engine.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <template id="frontend_homepage.js" name="Frontend: JS" inherit_id="frontend_base.js">
+            <xpath expr="." position="inside">
+                <script type="text/javascript" src="/frontend_homepage/static/js/geocharts.js"></script>
+            </xpath>
+        </template>
+
+    </data>
+</openerp>
