+++ main/local_modules/main/__openerp__.py
@@ -59,8 +59,6 @@ Contributors
         'geoengine_base_geolocalize',
         # website
         'website_blog_mgmt',
-        # odoo sentry
-        'odoo_sentry',
         # local modules
         'res_partner_email_validation',
         'res_partner_email_validation_missing_details',
